+++ jannovar-htsjdk/src/main/java/de/charite/compbio/jannovar/htsjdk/VariantContextAnnotator.java
@@ -2,6 +2,7 @@ package de.charite.compbio.jannovar.htsjdk;
 
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -12,6 +13,7 @@ import org.slf4j.LoggerFactory;
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
 
 import de.charite.compbio.jannovar.annotation.Annotation;
 import de.charite.compbio.jannovar.annotation.AnnotationMessage;
@@ -28,6 +30,7 @@ import de.charite.compbio.jannovar.reference.Strand;
 import de.charite.compbio.jannovar.reference.TranscriptModel;
 import htsjdk.variant.variantcontext.Allele;
 import htsjdk.variant.variantcontext.VariantContext;
+import htsjdk.variant.variantcontext.VariantContextBuilder;
 
 /**
  * Helper class for generating {@link VariantAnnotations} objects from {@link VariantContext}s.
@@ -59,6 +62,15 @@ public final class VariantContextAnnotator {
 		/** whether or not to perform shifting towards the 3' end of the transcript (defaults to <code>true</code>) */
 		private final boolean nt3PrimeShifting;
 
+		/** Whether or not off-target filter is enabled */
+		private boolean offTargetFilterEnabled;
+
+		/** Whether or not UTR counts as off-target */
+		private boolean offTargetFilterUtrIsOffTarget;
+
+		/** Whether or not non-consensus splice region counts as off-target */
+		private boolean offTargetFilterIntronicSpliceIsOffTarget;
+
 		/**
 		 * Constructor
 		 */
@@ -66,6 +78,9 @@ public final class VariantContextAnnotator {
 			oneAnnotationOnly = true;
 			escapeAnnField = true;
 			nt3PrimeShifting = true;
+			offTargetFilterEnabled = false;
+			offTargetFilterUtrIsOffTarget = false;
+			offTargetFilterIntronicSpliceIsOffTarget = false;
 		}
 
 		/**
@@ -80,11 +95,22 @@ public final class VariantContextAnnotator {
 		 * @param nt3PrimeShifting
 		 *            whether or not to perform shifting towards the 3' end of the transcript (defaults to
 		 *            <code>true</code>)
+		 * @param offTargetFilterEnabled
+		 *            whether or not off target filter application is abled
+		 * @param offTargetFilterUtrIsOffTarget
+		 *            whether or not to count UTR as off-target
+		 * @param offTargetFilterIntronicSpliceIsOffTarget
+		 *            whether or not to to count non-consensus intronic splicing as off-target
 		 */
-		public Options(boolean oneAnnotationOnly, boolean escapeAnnField, boolean nt3PrimeShifting) {
+		public Options(boolean oneAnnotationOnly, boolean escapeAnnField, boolean nt3PrimeShifting,
+				boolean offTargetFilterEnabled, boolean offTargetFilterUtrIsOffTarget,
+				boolean offTargetFilterIntronicSpliceIsOffTarget) {
 			this.oneAnnotationOnly = oneAnnotationOnly;
 			this.escapeAnnField = escapeAnnField;
 			this.nt3PrimeShifting = nt3PrimeShifting;
+			this.offTargetFilterEnabled = offTargetFilterEnabled;
+			this.offTargetFilterUtrIsOffTarget = offTargetFilterUtrIsOffTarget;
+			this.offTargetFilterIntronicSpliceIsOffTarget = offTargetFilterIntronicSpliceIsOffTarget;
 		}
 
 		/**
@@ -106,6 +132,18 @@ public final class VariantContextAnnotator {
 		 */
 		public boolean isNt3PrimeShifting() {
 			return nt3PrimeShifting;
+		}
+
+		public boolean isOffTargetFilterEnabled() {
+			return offTargetFilterEnabled;
+		}
+
+		public boolean isOffTargetFilterUtrIsOffTarget() {
+			return offTargetFilterUtrIsOffTarget;
+		}
+
+		public boolean isOffTargetFilterIntronicSpliceIsOffTarget() {
+			return offTargetFilterIntronicSpliceIsOffTarget;
 		}
 
 	}
@@ -231,6 +269,22 @@ public final class VariantContextAnnotator {
 	}
 
 	/**
+	 * Annotate variant <code>vc</code> and return annoated variant
+	 * 
+	 * @param vc
+	 *            {@link VariantContext} to annotate
+	 */
+	public VariantContext annotateVariantContext(VariantContext vc) {
+		try {
+			vc = applyAnnotations(vc, buildAnnotations(vc));
+		} catch (InvalidCoordinatesException e) {
+			putErrorAnnotation(vc, ImmutableSet.of(e.getAnnotationMessage()));
+		}
+		vc.getCommonInfo().removeAttribute(""); // remove leading/trailing comma
+		return vc;
+	}
+
+	/**
 	 * Given a {@link VariantContext}, generate one {@link VariantAnnotations} for each alternative allele.
 	 *
 	 * Note that in the case of an exception being thrown, you have to add an error annotation yourself to the
@@ -276,6 +330,9 @@ public final class VariantContextAnnotator {
 	 * @return modified <code>vc</code>
 	 */
 	public VariantContext applyAnnotations(VariantContext vc, List<VariantAnnotations> annos) {
+		// Whether or not variant is off-target in all annotations
+		boolean offTargetInAll = false;
+
 		ArrayList<String> annotations = new ArrayList<String>();
 		for (int alleleID = 0; alleleID < vc.getAlternateAlleles().size(); ++alleleID) {
 			if (!annos.get(alleleID).getAnnotations().isEmpty()) {
@@ -286,6 +343,12 @@ public final class VariantContextAnnotator {
 						break;
 				}
 			}
+		}
+
+		if (options.isOffTargetFilterEnabled() && (offTargetInAll && !annotations.isEmpty())) {
+			Set<String> filters = new HashSet<>(vc.getFilters());
+			// ASDF
+			vc = new VariantContextBuilder(vc).filters(filters).make();
 		}
 
 		// If a VC builder is used before the attributes can be unmodifiable.